pyproject.toml
[tool.poetry]
name = "drf-filter"
version = "0.1.0"
description = "Examples of django-filter in django rest framework"
authors = ["Khemmatat Theanvanichpant <khemmatat.thean@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-extensions = "^3.1.5"
django-filter = "^21.1"
drf-spectacular = "^0.21.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-clarity = "^1.0.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120
extend_skip_glob = ["*/migrations/*"]

[tool.black]
line-length = 120
target-version = ["py39"]
extend-exclude = '''
/(
    .+\/migrations
)/
'''
